# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "RT Backend API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "components": {
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "coordinator",
              "tutor",
              "guardian"
            ]
          },
          "rut": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          },
          "chargeEmail": {
            "type": "string",
            "nullable": true
          },
          "institutionId": {
            "type": "integer",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "email",
          "role",
          "createdAt",
          "updatedAt"
        ]
      },
      "UserInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "User full name (required)"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "User email (must be unique, required)"
          },
          "role": {
            "type": "string",
            "enum": [
              "admin",
              "coordinator",
              "tutor",
              "guardian"
            ],
            "description": "User role (required)"
          },
          "rut": {
            "type": "string",
            "description": "RUT in format XX.XXX.XXX-K (required, used to generate initial password)"
          },
          "phone": {
            "type": "string",
            "nullable": true,
            "description": "Phone number (optional)"
          },
          "address": {
            "type": "string",
            "nullable": true,
            "description": "Address (optional)"
          },
          "chargeEmail": {
            "type": "string",
            "format": "email",
            "nullable": true,
            "description": "Charge email (optional)"
          },
          "institutionId": {
            "type": "integer",
            "nullable": true,
            "description": "Institution ID (required for non-admin users, inferred for coordinators)"
          }
        },
        "required": [
          "name",
          "email",
          "role",
          "rut"
        ]
      },
      "Institution": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "name",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "UserWithInstitution": {
        "allOf": [
          {
            "$ref": "#/components/schemas/User"
          },
          {
            "type": "object",
            "properties": {
              "Institution": {
                "$ref": "#/components/schemas/Institution"
              }
            }
          }
        ]
      },
      "UserBankAccount": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "userId": {
            "type": "integer"
          },
          "bankName": {
            "type": "string"
          },
          "accountType": {
            "type": "string",
            "enum": [
              "AHORRO",
              "CORRIENTE",
              "VISTA"
            ]
          },
          "accountNumber": {
            "type": "string"
          },
          "accountEmail": {
            "type": "string",
            "format": "email"
          },
          "accountName": {
            "type": "string"
          },
          "rut": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "Student": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "guardianId": {
            "type": "integer"
          },
          "institutionId": {
            "type": "integer"
          },
          "isActive": {
            "type": "boolean"
          }
        }
      },
...
      "GuardianTutor": {
        "type": "object",
        "properties": {
          "guardianId": {
            "type": "integer"
          },
          "tutorId": {
            "type": "integer"
          },
          "institutionId": {
            "type": "integer"
          },
          "active": {
            "type": "boolean"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "TutorLink": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/GuardianTutor"
          },
          {
            "type": "object",
            "properties": {
              "Guardian": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        ]
      },
      "GuardianLink": {
        "type": "object",
        "allOf": [
          {
            "$ref": "#/components/schemas/GuardianTutor"
          },
          {
            "type": "object",
            "properties": {
              "Tutor": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        ]
      },
      "UserDetail": {
        "allOf": [
          {
            "$ref": "#/components/schemas/User"
          },
          {
            "type": "object",
            "properties": {
              "Institution": {
                "$ref": "#/components/schemas/Institution",
                "nullable": true
              },
              "BankAccount": {
                "$ref": "#/components/schemas/UserBankAccount",
                "nullable": true
              },
              "Students": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/Student"
                }
              },
              "TutorLinks": {
                "type": "array",
                "items": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/GuardianTutor"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "Guardian": {
                          "$ref": "#/components/schemas/User"
                        }
                      }
                    }
                  ]
                }
              },
              "GuardianLinks": {
                "type": "array",
                "items": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/GuardianTutor"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "Tutor": {
                          "$ref": "#/components/schemas/User"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        ]
      },
      "UserBankAccountInput": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "integer"
          },
          "bankName": {
            "type": "string"
          },
          "accountType": {
            "type": "string"
          },
          "accountNumber": {
            "type": "string"
          },
          "rut": {
            "type": "string"
          },
          "accountEmail": {
            "type": "string",
            "format": "email"
          },
          "accountName": {
            "type": "string"
          }
        },
        "required": [
          "bankName",
          "accountType",
          "accountNumber",
          "rut",
          "accountEmail",
          "accountName"
        ]
      },
...
      "CreateUserWithBankAccountInput": {
        "allOf": [
          {
            "$ref": "#/components/schemas/UserInput"
          },
          {
            "type": "object",
            "properties": {
              "BankAccount": {
                "allOf": [
                  {
                    "$ref": "#/components/schemas/UserBankAccountInput"
                  }
                ],
                "nullable": true
              }
            }
          }
        ]
      },
      "UserByIdResponse": {
        "description": "UserDetail with optional coordinatorProfitShare field. coordinatorProfitShare is only present when user role is coordinator.",
        "allOf": [
          {
            "$ref": "#/components/schemas/UserDetail"
          },
          {
            "type": "object",
            "properties": {
              "coordinatorProfitShare": {
                "type": "number",
                "nullable": true,
                "description": "Profit share percentage for coordinator users. Only present when user role is coordinator."
              }
            }
          }
        ]
      },
      "EditUserPersonalInformationInput": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "rut": {
            "type": "string",
            "nullable": true
          },
          "phone": {
            "type": "string",
            "nullable": true
          },
          "chargeEmail": {
            "type": "string",
            "nullable": true
          },
          "address": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "name",
          "phone",
          "rut"
        ]
      },
      "ChangePasswordRequest": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string"
          },
          "newPassword": {
            "type": "string"
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
...
    }
  },
  "paths": {
...
    "/users": {
      "get": {
        "summary": "List users",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "role",
            "schema": {
              "type": "string",
              "enum": [
                "admin",
                "coordinator",
                "tutor",
                "guardian"
              ]
            },
            "description": "Filter by user role"
          },
          {
            "in": "query",
            "name": "institutionId",
            "schema": {
              "type": "integer"
            },
            "description": "Filter by institution id"
          },
          {
            "in": "query",
            "name": "nameOrEmail",
            "schema": {
              "type": "string"
            },
            "description": "Case-insensitive search in name or email"
          },
          {
            "in": "query",
            "name": "sendInactive",
            "schema": {
              "type": "boolean"
            },
            "description": "If false, only active users are returned. If true or omitted, returns all."
          },
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            },
            "description": "Page number (1-based)"
          },
          {
            "in": "query",
            "name": "pageSize",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 10
            },
            "description": "Items per page"
          }
        ],
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/UserWithInstitution"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a user",
        "description": "Create a new user in the system.\n- Admins must provide the institution ID\n- Coordinators automatically use their own institution\n- Initial password is set to the RUT number without the verifying digit\n- Email must be unique (database constraint)\n- Phone, address, and chargeEmail are optional\n",
        "tags": [
          "Users"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserWithBankAccountInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "400": {
            "description": "Invalid input or validation error (missing required fields, invalid email format, or database constraint violation)"
          },
          "403": {
            "description": "Forbidden - insufficient permissions"
          }
        }
      }
    },
    "/users/{id}": {
      "delete": {
        "summary": "Delete a user by ID",
        "description": "Soft delete a user by marking them as inactive (isActive = false).\n- Admins and coordinators can delete users\n- Coordinators cannot delete admin or coordinator users\n",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID"
          }
        ],
        "responses": {
          "204": {
            "description": "User deleted successfully (soft delete)"
          },
          "403": {
            "description": "Forbidden - user lacks permission to delete this user"
          },
          "404": {
            "description": "User not found"
          }
        }
      },
      "get": {
        "summary": "Get a user by ID",
        "description": "Returns user details.\n- For guardians/tutors, only active students are returned\n",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "User found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserByIdResponse"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/users/{id}/bank-account": {
      "patch": {
        "summary": "Edit a user's bank account information",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserBankAccountInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated bank account information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserBankAccount"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/users/{id}/personal-information": {
      "patch": {
        "summary": "Edit a user's personal information",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditUserPersonalInformationInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Updated user information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/User"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/users/{id}/change-password": {
      "patch": {
        "summary": "Change a user's password",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "message": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Current password is incorrect"
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/users/{id}/tutor-links": {
      "get": {
        "summary": "Get tutor links for a user",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Tutor links retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TutorLink"
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    },
    "/users/{id}/guardian-links": {
      "get": {
        "summary": "Get guardian links for a user",
        "tags": [
          "Users"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "User ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Guardian links retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GuardianLink"
                  }
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        }
      }
    }
  },
...
}